requirements.txt
faiss-cpu
numpy
sentence-transformers
python-docx
Flask
requests
groq
python-dotenv